DOSIS-1700 Insert organisations into repo for existing tests
