+++ tests/service_automation/tests/json_files/Organisation/1481a74c-8dd3-4b5b-bca5-dd1701f25b09.json
@@ -1,22 +1,22 @@
 {
-  "id": "00000000-bab3-4baf-92da-0c77df9363a6",
+  "id": "1481a74c-8dd3-4b5b-bca5-dd1701f25b09",
   "createdBy": "ROBOT",
   "createdDateTime": "2025-06-27T12:46:36.111296Z",
   "modifiedBy": "ROBOT",
   "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
-  "identifier_ODS_ODSCode": "H82028",
+  "identifier_ODS_ODSCode": "N00081063",
   "active": true,
-  "name": "Test Organisation",
+  "name": "GP: York Road Group Practice - Ellesmere Port",
   "telecom": null,
   "type": "GP Practice",
   "endpoints": [
     {
-      "id": "a85c53bd-3093-47a1-a49b-adb597a42329",
+      "id": "423a9da5-72ab-434c-a48f-d1d941b8f653",
       "createdBy": "ROBOT",
       "createdDateTime": "2025-06-27T12:46:36.111296Z",
       "modifiedBy": "ROBOT",
       "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
-      "identifier_oldDoS_id": 9222,
+      "identifier_oldDoS_id": 4011,
       "status": "active",
       "connectionType": "itk",
       "name": null,
@@ -24,18 +24,18 @@
       "description": "Primary",
       "payloadType": "urn:nhs-itk:interaction:primaryGeneralPractitionerRecipientNHS111CDADocument-v2-0",
       "address": "https://dummy-itk-endpoint.nhs.uk",
-      "managedByOrganisation": "00000000-bab3-4baf-92da-0c77df9363a6",
+      "managedByOrganisation": "1481a74c-8dd3-4b5b-bca5-dd1701f25b09",
       "service": null,
       "order": 1,
       "isCompressionEnabled": false
     },
     {
-      "id": "44c8c179-e180-400d-8ce3-3d4eb52535b4",
+      "id": "e9d94f29-3330-4fb4-ae92-4a25accd074d",
       "createdBy": "ROBOT",
       "createdDateTime": "2025-06-27T12:46:36.111296Z",
       "modifiedBy": "ROBOT",
       "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
-      "identifier_oldDoS_id": 9223,
+      "identifier_oldDoS_id": 4012,
       "status": "active",
       "connectionType": "itk",
       "name": null,
@@ -43,45 +43,45 @@
       "description": "Copy",
       "payloadType": "urn:nhs-itk:interaction:copyRecipientNHS111CDADocument-v2-0",
       "address": "https://dummy-itk-endpoint.nhs.uk",
-      "managedByOrganisation": "00000000-bab3-4baf-92da-0c77df9363a6",
+      "managedByOrganisation": "1481a74c-8dd3-4b5b-bca5-dd1701f25b09",
       "service": null,
       "order": 1,
       "isCompressionEnabled": false
     },
     {
-      "id": "5f32255f-6d66-4491-8bf8-a9ab43db2228",
+      "id": "3f296234-26f7-4ac4-9ef1-8c7028200949",
       "createdBy": "ROBOT",
       "createdDateTime": "2025-06-27T12:46:36.111296Z",
       "modifiedBy": "ROBOT",
       "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
-      "identifier_oldDoS_id": 370382,
+      "identifier_oldDoS_id": 293398,
       "status": "active",
       "connectionType": "email",
       "name": null,
       "payloadMimeType": "application/pdf",
-      "description": "Copy",
-      "payloadType": "urn:nhs-itk:interaction:copyRecipientNHS111CDADocument-v2-0",
+      "description": "Primary",
+      "payloadType": "urn:nhs-itk:interaction:primaryGeneralPractitionerRecipientNHS111CDADocument-v2-0",
       "address": "dummy-endpoint-email@nhs.net",
-      "managedByOrganisation": "00000000-bab3-4baf-92da-0c77df9363a6",
+      "managedByOrganisation": "1481a74c-8dd3-4b5b-bca5-dd1701f25b09",
       "service": null,
       "order": 2,
       "isCompressionEnabled": false
     },
     {
-      "id": "59055b9b-d12f-4d6f-bf39-9fe1eaff7f2c",
+      "id": "ebaaea78-f8dc-48e5-aa3b-3537475203b7",
       "createdBy": "ROBOT",
       "createdDateTime": "2025-06-27T12:46:36.111296Z",
       "modifiedBy": "ROBOT",
       "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
-      "identifier_oldDoS_id": 370383,
+      "identifier_oldDoS_id": 293399,
       "status": "active",
       "connectionType": "email",
       "name": null,
       "payloadMimeType": "application/pdf",
-      "description": "Primary",
-      "payloadType": "urn:nhs-itk:interaction:primaryGeneralPractitionerRecipientNHS111CDADocument-v2-0",
+      "description": "Copy",
+      "payloadType": "urn:nhs-itk:interaction:copyRecipientNHS111CDADocument-v2-0",
       "address": "dummy-endpoint-email@nhs.net",
-      "managedByOrganisation": "00000000-bab3-4baf-92da-0c77df9363a6",
+      "managedByOrganisation": "1481a74c-8dd3-4b5b-bca5-dd1701f25b09",
       "service": null,
       "order": 2,
       "isCompressionEnabled": false
