+++ tests/service_automation/tests/json_files/Organisation/315809be-33ab-46b2-850c-f81e3ec0366b.json
@@ -1,0 +1,90 @@
+{
+  "id": "315809be-33ab-46b2-850c-f81e3ec0366b",
+  "createdBy": "ROBOT",
+  "createdDateTime": "2025-06-27T12:46:36.111296Z",
+  "modifiedBy": "ROBOT",
+  "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+  "identifier_ODS_ODSCode": "E8104",
+  "active": true,
+  "name": "GP - Sundon Medical Centre - 142-144 Sundon Park Road - Luton",
+  "telecom": null,
+  "type": "GP Practice",
+  "endpoints": [
+    {
+      "id": "2437aa9c-e711-487b-918e-e44851c90a46",
+      "createdBy": "ROBOT",
+      "createdDateTime": "2025-06-27T12:46:36.111296Z",
+      "modifiedBy": "ROBOT",
+      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+      "identifier_oldDoS_id": 330831,
+      "status": "active",
+      "connectionType": "email",
+      "name": null,
+      "payloadMimeType": "application/pdf",
+      "description": "Primary",
+      "payloadType": "urn:nhs-itk:interaction:primaryGeneralPractitionerRecipientNHS111CDADocument-v2-0",
+      "address": "dummy-endpoint-email@nhs.net",
+      "managedByOrganisation": "315809be-33ab-46b2-850c-f81e3ec0366b",
+      "service": null,
+      "order": 2,
+      "isCompressionEnabled": false
+    },
+    {
+      "id": "677a8baa-f578-4e57-b582-8c1e8891337d",
+      "createdBy": "ROBOT",
+      "createdDateTime": "2025-06-27T12:46:36.111296Z",
+      "modifiedBy": "ROBOT",
+      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+      "identifier_oldDoS_id": 330832,
+      "status": "active",
+      "connectionType": "email",
+      "name": null,
+      "payloadMimeType": "application/pdf",
+      "description": "Copy",
+      "payloadType": "urn:nhs-itk:interaction:copyRecipientNHS111CDADocument-v2-0",
+      "address": "dummy-endpoint-email@nhs.net",
+      "managedByOrganisation": "315809be-33ab-46b2-850c-f81e3ec0366b",
+      "service": null,
+      "order": 2,
+      "isCompressionEnabled": false
+    },
+    {
+      "id": "4b3b64ec-9501-401c-a5aa-3ce57f9168e7",
+      "createdBy": "ROBOT",
+      "createdDateTime": "2025-06-27T12:46:36.111296Z",
+      "modifiedBy": "ROBOT",
+      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+      "identifier_oldDoS_id": 185079,
+      "status": "active",
+      "connectionType": "itk",
+      "name": null,
+      "payloadMimeType": "application/hl7-cda+xml",
+      "description": "Copy",
+      "payloadType": "urn:nhs-itk:interaction:copyRecipientNHS111CDADocument-v2-0",
+      "address": "https://dummy-itk-endpoint.nhs.uk",
+      "managedByOrganisation": "315809be-33ab-46b2-850c-f81e3ec0366b",
+      "service": null,
+      "order": 1,
+      "isCompressionEnabled": false
+    },
+    {
+      "id": "fbf0a375-e86a-4698-9d67-f5aa3f7325ec",
+      "createdBy": "ROBOT",
+      "createdDateTime": "2025-06-27T12:46:36.111296Z",
+      "modifiedBy": "ROBOT",
+      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+      "identifier_oldDoS_id": 184531,
+      "status": "active",
+      "connectionType": "itk",
+      "name": null,
+      "payloadMimeType": "application/hl7-cda+xml",
+      "description": "Primary",
+      "payloadType": "urn:nhs-itk:interaction:primaryGeneralPractitionerRecipientNHS111CDADocument-v2-0",
+      "address": "https://dummy-itk-endpoint.nhs.uk",
+      "managedByOrganisation": "315809be-33ab-46b2-850c-f81e3ec0366b",
+      "service": null,
+      "order": 1,
+      "isCompressionEnabled": false
+    }
+  ]
+}
