+++ tests/service_automation/tests/json_files/Organisation/36fce427-0f31-4a4e-903f-74dcf9e63cfd.json
@@ -1,79 +1,22 @@
 {
-  "id": "11111111-bab3-4baf-92da-0c77df9363a6",
+  "id": "36fce427-0f31-4a4e-903f-74dcf9e63cfd",
   "createdBy": "ROBOT",
   "createdDateTime": "2025-06-27T12:46:36.111296Z",
   "modifiedBy": "ROBOT",
   "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
-  "identifier_ODS_ODSCode": "H82028",
+  "identifier_ODS_ODSCode": "M00081046",
   "active": true,
-  "name": "Test Organisation",
+  "name": "Abbottswood Medical Practice, Pershore, Worcestershire",
   "telecom": null,
   "type": "GP Practice",
   "endpoints": [
     {
-      "id": "a85c53bd-3093-47a1-a49b-adb597a42329",
+      "id": "d449e3b8-eba2-43ec-9ee5-aee85387b53d",
       "createdBy": "ROBOT",
       "createdDateTime": "2025-06-27T12:46:36.111296Z",
       "modifiedBy": "ROBOT",
       "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
-      "identifier_oldDoS_id": 9222,
-      "status": "active",
-      "connectionType": "itk",
-      "name": null,
-      "payloadMimeType": "application/hl7-cda+xml",
-      "description": "Primary",
-      "payloadType": "urn:nhs-itk:interaction:primaryGeneralPractitionerRecipientNHS111CDADocument-v2-0",
-      "address": "https://dummy-itk-endpoint.nhs.uk",
-      "managedByOrganisation": "11111111-bab3-4baf-92da-0c77df9363a6",
-      "service": null,
-      "order": 1,
-      "isCompressionEnabled": false
-    },
-    {
-      "id": "44c8c179-e180-400d-8ce3-3d4eb52535b4",
-      "createdBy": "ROBOT",
-      "createdDateTime": "2025-06-27T12:46:36.111296Z",
-      "modifiedBy": "ROBOT",
-      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
-      "identifier_oldDoS_id": 9223,
-      "status": "active",
-      "connectionType": "itk",
-      "name": null,
-      "payloadMimeType": "application/hl7-cda+xml",
-      "description": "Copy",
-      "payloadType": "urn:nhs-itk:interaction:copyRecipientNHS111CDADocument-v2-0",
-      "address": "https://dummy-itk-endpoint.nhs.uk",
-      "managedByOrganisation": "11111111-bab3-4baf-92da-0c77df9363a6",
-      "service": null,
-      "order": 1,
-      "isCompressionEnabled": false
-    },
-    {
-      "id": "5f32255f-6d66-4491-8bf8-a9ab43db2228",
-      "createdBy": "ROBOT",
-      "createdDateTime": "2025-06-27T12:46:36.111296Z",
-      "modifiedBy": "ROBOT",
-      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
-      "identifier_oldDoS_id": 370382,
-      "status": "active",
-      "connectionType": "email",
-      "name": null,
-      "payloadMimeType": "application/pdf",
-      "description": "Copy",
-      "payloadType": "urn:nhs-itk:interaction:copyRecipientNHS111CDADocument-v2-0",
-      "address": "dummy-endpoint-email@nhs.net",
-      "managedByOrganisation": "11111111-bab3-4baf-92da-0c77df9363a6",
-      "service": null,
-      "order": 2,
-      "isCompressionEnabled": false
-    },
-    {
-      "id": "59055b9b-d12f-4d6f-bf39-9fe1eaff7f2c",
-      "createdBy": "ROBOT",
-      "createdDateTime": "2025-06-27T12:46:36.111296Z",
-      "modifiedBy": "ROBOT",
-      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
-      "identifier_oldDoS_id": 370383,
+      "identifier_oldDoS_id": 212417,
       "status": "active",
       "connectionType": "email",
       "name": null,
@@ -81,9 +24,66 @@
       "description": "Primary",
       "payloadType": "urn:nhs-itk:interaction:primaryGeneralPractitionerRecipientNHS111CDADocument-v2-0",
       "address": "dummy-endpoint-email@nhs.net",
-      "managedByOrganisation": "11111111-bab3-4baf-92da-0c77df9363a6",
+      "managedByOrganisation": "36fce427-0f31-4a4e-903f-74dcf9e63cfd",
       "service": null,
       "order": 2,
+      "isCompressionEnabled": false
+    },
+    {
+      "id": "6b2077f2-23b3-4d65-8a72-cb58b26e42d1",
+      "createdBy": "ROBOT",
+      "createdDateTime": "2025-06-27T12:46:36.111296Z",
+      "modifiedBy": "ROBOT",
+      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+      "identifier_oldDoS_id": 212415,
+      "status": "active",
+      "connectionType": "email",
+      "name": null,
+      "payloadMimeType": "application/pdf",
+      "description": "Copy",
+      "payloadType": "urn:nhs-itk:interaction:copyRecipientNHS111CDADocument-v2-0",
+      "address": "dummy-endpoint-email@nhs.net",
+      "managedByOrganisation": "36fce427-0f31-4a4e-903f-74dcf9e63cfd",
+      "service": null,
+      "order": 2,
+      "isCompressionEnabled": false
+    },
+    {
+      "id": "93e43c4a-a199-4927-97b7-34200b59955e",
+      "createdBy": "ROBOT",
+      "createdDateTime": "2025-06-27T12:46:36.111296Z",
+      "modifiedBy": "ROBOT",
+      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+      "identifier_oldDoS_id": 212413,
+      "status": "active",
+      "connectionType": "itk",
+      "name": null,
+      "payloadMimeType": "application/hl7-cda+xml",
+      "description": "Copy",
+      "payloadType": "urn:nhs-itk:interaction:copyRecipientNHS111CDADocument-v2-0",
+      "address": "https://dummy-itk-endpoint.nhs.uk",
+      "managedByOrganisation": "36fce427-0f31-4a4e-903f-74dcf9e63cfd",
+      "service": null,
+      "order": 1,
+      "isCompressionEnabled": false
+    },
+    {
+      "id": "1a0b103c-77b6-472a-91c9-fd0e4895b85c",
+      "createdBy": "ROBOT",
+      "createdDateTime": "2025-06-27T12:46:36.111296Z",
+      "modifiedBy": "ROBOT",
+      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+      "identifier_oldDoS_id": 214755,
+      "status": "active",
+      "connectionType": "itk",
+      "name": null,
+      "payloadMimeType": "application/hl7-cda+xml",
+      "description": "Primary",
+      "payloadType": "urn:nhs-itk:interaction:primaryGeneralPractitionerRecipientNHS111CDADocument-v2-0",
+      "address": "https://dummy-itk-endpoint.nhs.uk",
+      "managedByOrganisation": "36fce427-0f31-4a4e-903f-74dcf9e63cfd",
+      "service": null,
+      "order": 1,
       "isCompressionEnabled": false
     }
   ]
