+++ tests/service_automation/tests/json_files/Organisation/377dbc2b-b2f9-4d14-9cc5-5360752b3a6c.json
@@ -1,0 +1,13 @@
+{
+  "id": "377dbc2b-b2f9-4d14-9cc5-5360752b3a6c",
+  "createdBy": "ROBOT",
+  "createdDateTime": "2025-06-27T12:46:36.111296Z",
+  "modifiedBy": "ROBOT",
+  "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+  "identifier_ODS_ODSCode": "F00081015",
+  "active": true,
+  "name": "GP: Crocus Medical Practice Saffron Walden West Essex",
+  "telecom": null,
+  "type": "GP Practice",
+  "endpoints": []
+}
