+++ tests/service_automation/tests/json_files/Organisation/dbb206d6-6cf7-43a4-b910-10fdc14a3cb6.json
@@ -1,0 +1,90 @@
+{
+  "id": "dbb206d6-6cf7-43a4-b910-10fdc14a3cb6",
+  "createdBy": "ROBOT",
+  "createdDateTime": "2025-06-27T12:46:36.111296Z",
+  "modifiedBy": "ROBOT",
+  "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+  "identifier_ODS_ODSCode": "H82028",
+  "active": true,
+  "name": "Test Organisation",
+  "telecom": null,
+  "type": "GP Practice",
+  "endpoints": [
+    {
+      "id": "a85c53bd-3093-47a1-a49b-adb597a42329",
+      "createdBy": "ROBOT",
+      "createdDateTime": "2025-06-27T12:46:36.111296Z",
+      "modifiedBy": "ROBOT",
+      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+      "identifier_oldDoS_id": 9222,
+      "status": "active",
+      "connectionType": "itk",
+      "name": null,
+      "payloadMimeType": "application/hl7-cda+xml",
+      "description": "Primary",
+      "payloadType": "urn:nhs-itk:interaction:primaryGeneralPractitionerRecipientNHS111CDADocument-v2-0",
+      "address": "https://dummy-itk-endpoint.nhs.uk",
+      "managedByOrganisation": "dbb206d6-6cf7-43a4-b910-10fdc14a3cb6",
+      "service": null,
+      "order": 1,
+      "isCompressionEnabled": false
+    },
+    {
+      "id": "44c8c179-e180-400d-8ce3-3d4eb52535b4",
+      "createdBy": "ROBOT",
+      "createdDateTime": "2025-06-27T12:46:36.111296Z",
+      "modifiedBy": "ROBOT",
+      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+      "identifier_oldDoS_id": 9223,
+      "status": "active",
+      "connectionType": "itk",
+      "name": null,
+      "payloadMimeType": "application/hl7-cda+xml",
+      "description": "Copy",
+      "payloadType": "urn:nhs-itk:interaction:copyRecipientNHS111CDADocument-v2-0",
+      "address": "https://dummy-itk-endpoint.nhs.uk",
+      "managedByOrganisation": "dbb206d6-6cf7-43a4-b910-10fdc14a3cb6",
+      "service": null,
+      "order": 1,
+      "isCompressionEnabled": false
+    },
+    {
+      "id": "5f32255f-6d66-4491-8bf8-a9ab43db2228",
+      "createdBy": "ROBOT",
+      "createdDateTime": "2025-06-27T12:46:36.111296Z",
+      "modifiedBy": "ROBOT",
+      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+      "identifier_oldDoS_id": 370382,
+      "status": "active",
+      "connectionType": "email",
+      "name": null,
+      "payloadMimeType": "application/pdf",
+      "description": "Copy",
+      "payloadType": "urn:nhs-itk:interaction:copyRecipientNHS111CDADocument-v2-0",
+      "address": "dummy-endpoint-email@nhs.net",
+      "managedByOrganisation": "dbb206d6-6cf7-43a4-b910-10fdc14a3cb6",
+      "service": null,
+      "order": 2,
+      "isCompressionEnabled": false
+    },
+    {
+      "id": "59055b9b-d12f-4d6f-bf39-9fe1eaff7f2c",
+      "createdBy": "ROBOT",
+      "createdDateTime": "2025-06-27T12:46:36.111296Z",
+      "modifiedBy": "ROBOT",
+      "modifiedDateTime": "2025-06-27T12:46:36.111296Z",
+      "identifier_oldDoS_id": 370383,
+      "status": "active",
+      "connectionType": "email",
+      "name": null,
+      "payloadMimeType": "application/pdf",
+      "description": "Primary",
+      "payloadType": "urn:nhs-itk:interaction:primaryGeneralPractitionerRecipientNHS111CDADocument-v2-0",
+      "address": "dummy-endpoint-email@nhs.net",
+      "managedByOrganisation": "dbb206d6-6cf7-43a4-b910-10fdc14a3cb6",
+      "service": null,
+      "order": 2,
+      "isCompressionEnabled": false
+    }
+  ]
+}
